Uszczelnij izolację danych graczy i obsadę hubów
Wdrożono rzeczywisty limit przetwarzania graczy w ticku z uczciwym wyborem partii i nadrabianiem czasu przez last_tick_at.

Dodano serwisy przypisań pracowników i wyliczania obsady hubów, zwalnianie przypisań po odejściu pracownika oraz przygotowanie integracji z logistyką.

Przeprowadzono przekrojowy audyt player_id i dodano kontrolę właściciela do operacji finansowych, HR, hubów, rurociągów, odwiertów, rynku, czatu, ochrony, kontraktów B2B i sekcji ticka. Mutacje końcowe sprawdzają również stan rekordu i wynik rowCount, co ogranicza wyścigi oraz podwójne rozliczenia.

Rozszerzono testy regresyjne, integracyjne i MySQL o izolację dyrektorów bankowych, ofert rynkowych, zadań technicznych, hubów, rurociągów, przypisań pracowników i limit PlayersModule. Zaktualizowano AGENTS.md oraz GAME_README.md.

Weryfikacja: Unit 91/91, Integration 454/454, MySqlIntegration 224/224, PHPStan dla zmienionego zakresu bez błędów, PHP lint 77/77, encoding 1912/1912 oraz git diff --check.

+++ src/Well/ActionsTrait.php
@@ -233,7 +233,7 @@ trait WellActionsTrait
     }
 
     private function logEvent(int $wellId, int $playerId, string $type, float $cost,
-                               string $desc, ?int $condBefore = null, ?int $condAfter = null): void
+                               string $desc, ?float $condBefore = null, ?float $condAfter = null): void
     {
         $this->db->prepare("
             INSERT INTO well_events (well_id, player_id, event_type, cost, description,
